requirements.txt
# Requirements dla Python 3.11+ (włącznie z 3.13)
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
lxml==5.3.0
Pillow==11.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0
openpyxl==3.1.5
pandas==2.2.3